chore: unescape GitHub Actions syntax
In order to be able to add GitHub Actions syntax in a action description, we need to escape it because GitHub will might otherwise fail as it will try to evaluate it. In the docs we want to show the unescaped values to the user again.

Escaped example: `\$\{\{ github.token \}\}`
unescaped example (in docs): ${{ github.token }}