chore: resolve the commit through git instead of github.sha
In some cases the commit-sha action is triggered on one branch, but as part of the workflow another branch is checked out. In these scenarios the commit-sha action should resolve the actual HEAD commot from the current branch instead of the github.sha, which refers to the branch that the workflow was triggered for.

This is useful especially for workflows that are triggered through the issue_commit trigger for a pull request. In these scenarios you still want to execute the workflow under the context of the PR branch instead of the default branch as workflows triggered for the issue_commit trigger are always triggered for the default branch for security reasons.